Allow overriding the fixture names, so other fixtures can depend on them
pytest doesn't have an API (yet?) like pytest.mark.usefixture(name) for fixtures,
so they need to use funcargs with the matching names to declare dependencies.
Which is of course not possible when the names are unknown to the client,
since they have include a non-deterministic part, the layer's object id.